% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_duckdb.R
\name{duckdb_exists}
\alias{duckdb_exists}
\title{Check if DuckDB database exists}
\usage{
duckdb_exists()
}
\value{
TRUE if database exists, FALSE otherwise
}
\description{
Check if DuckDB database file exists in app_folder
}
