% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_athena.R
\name{get_all_related_concepts}
\alias{get_all_related_concepts}
\title{Get all related concepts (combined)}
\usage{
get_all_related_concepts(concept_id, vocabularies, existing_mappings)
}
\description{
Get all related concepts (combined)
}
