% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vocabularies.R
\name{get_concept_synonyms}
\alias{get_concept_synonyms}
\title{Get concept synonyms from concept_synonym table}
\usage{
get_concept_synonyms(concept_id, vocabularies)
}
\description{
Get concept synonyms from concept_synonym table
}
