% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_athena.R
\name{get_descendant_concepts}
\alias{get_descendant_concepts}
\title{Get descendant concepts from concept_ancestor}
\usage{
get_descendant_concepts(concept_id, vocabularies)
}
\description{
Get descendant concepts from concept_ancestor
}
