% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_athena.R
\name{get_related_concepts}
\alias{get_related_concepts}
\title{Get related concepts (maps to / mapped from) - all concepts without filtering}
\usage{
get_related_concepts(concept_id, vocabularies)
}
\description{
Get related concepts (maps to / mapped from) - all concepts without filtering
}
