% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_config.R
\name{get_vocab_folder}
\alias{get_vocab_folder}
\title{Get OHDSI Vocabularies folder path}
\usage{
get_vocab_folder()
}
\value{
Folder path or NULL if not configured
}
\description{
Get the configured OHDSI Vocabularies folder path
}
