% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the INDICATE Shiny Application}
\usage{
run_app(app_folder = NULL, ..., options = list())
}
\arguments{
\item{app_folder}{Path to the folder where application files will be stored.
If NULL (default), uses the user's home directory.}

\item{...}{Additional arguments passed to \code{\link[shiny]{shinyApp}}}

\item{options}{A named list of options to pass to \code{\link[shiny]{shinyApp}}.
By default, the browser is launched automatically.}
}
\value{
A Shiny app object
}
\description{
Launch the INDICATE Minimal Data Dictionary Explorer application
}
\examples{
\dontrun{
run_app()
run_app(app_folder = "~/Documents/my_indicate_data")
}
}
