% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{set_use_duckdb}
\alias{set_use_duckdb}
\title{Set DuckDB option status}
\usage{
set_use_duckdb(use_duckdb)
}
\arguments{
\item{use_duckdb}{Logical value}
}
\value{
TRUE if successful
}
\description{
Set whether DuckDB database should be used
}
