% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_database.R
\name{set_vocab_folder}
\alias{set_vocab_folder}
\title{Set OHDSI Vocabularies folder path}
\usage{
set_vocab_folder(path)
}
\arguments{
\item{path}{Folder path}
}
\value{
TRUE if successful
}
\description{
Save the OHDSI Vocabularies folder path
}
