% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_duckdb.R
\name{delete_duckdb_database}
\alias{delete_duckdb_database}
\title{Delete DuckDB database}
\usage{
delete_duckdb_database()
}
\value{
List with success status and message
}
\description{
Delete the DuckDB database file from app_folder
}
