% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_athena.R
\name{get_descendant_concepts}
\alias{get_descendant_concepts}
\title{Get all hierarchy concepts (ancestors and descendants) for a given concept}
\usage{
get_descendant_concepts(concept_id, vocabularies)
}
\description{
Get all hierarchy concepts (ancestors and descendants) for a given concept
}
