% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_duckdb.R
\name{load_vocabularies_from_duckdb}
\alias{load_vocabularies_from_duckdb}
\title{Load vocabularies from DuckDB}
\usage{
load_vocabularies_from_duckdb()
}
\value{
List with dplyr::tbl connections to concept, concept_relationship, and concept_ancestor tables
}
\description{
Load OHDSI vocabularies from DuckDB database using lazy connections
}
