% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the INDICATE Shiny Application}
\usage{
run_app(..., options = list(launch.browser = TRUE))
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny]{shinyApp}}}

\item{options}{A named list of options to pass to \code{\link[shiny]{shinyApp}}.
By default, the browser is launched automatically.}
}
\value{
A Shiny app object
}
\description{
Launch the INDICATE Minimal Data Dictionary Explorer application
}
\examples{
\dontrun{
run_app()
}
}
